Fund: Uso de mappers para todos los dtos con prisma v3.1 Leonardo
