+++ backend/src/api/services/mappers/index.ts
@@ -1,0 +1,5 @@
+export { toComprobanteResponseDto } from './comprobante.mapper';
+export { toCursoResponseDto } from './curso.mapper';
+export { toDatosFacturacionResponseDto } from './datosFacturacion.mapper';
+export { toDatosPersonalesResponseDto } from './datosPersonales.mapper';
+export { toInscripcionResponseDto, type PrismaInscripcionConRelaciones } from './inscripcion.mapper';
